+++ internal/discovery/glob_test.go
@@ -5,7 +5,6 @@ import (
 	"errors"
 	"os"
 	"path"
-	"path/filepath"
 	"regexp"
 	"strconv"
 	"strings"
@@ -15,6 +14,7 @@ import (
 	"github.com/stretchr/testify/require"
 
 	"github.com/cloudflare/pint/internal/discovery"
+	"github.com/cloudflare/pint/internal/git"
 	"github.com/cloudflare/pint/internal/parser"
 )
 
@@ -43,32 +43,32 @@ func TestGlobPathFinder(t *testing.T) {
 	testCases := []testCaseT{
 		{
 			files:  map[string]string{},
-			finder: discovery.NewGlobFinder([]string{"[]"}, discovery.NewPathFilter(nil, nil, nil)),
-			err:    filepath.ErrBadPattern.Error(),
+			finder: discovery.NewGlobFinder([]string{"[]"}, git.NewPathFilter(nil, nil, nil)),
+			err:    "failed to expand file path pattern []: syntax error in pattern",
 		},
 		{
 			files:  map[string]string{},
-			finder: discovery.NewGlobFinder([]string{"*"}, discovery.NewPathFilter(nil, nil, nil)),
+			finder: discovery.NewGlobFinder([]string{"*"}, git.NewPathFilter(nil, nil, nil)),
 			err:    "no matching files",
 		},
 		{
 			files:  map[string]string{},
-			finder: discovery.NewGlobFinder([]string{"*"}, discovery.NewPathFilter(nil, nil, nil)),
+			finder: discovery.NewGlobFinder([]string{"*"}, git.NewPathFilter(nil, nil, nil)),
 			err:    "no matching files",
 		},
 		{
 			files:  map[string]string{},
-			finder: discovery.NewGlobFinder([]string{"foo/*"}, discovery.NewPathFilter(nil, nil, nil)),
+			finder: discovery.NewGlobFinder([]string{"foo/*"}, git.NewPathFilter(nil, nil, nil)),
 			err:    "no matching files",
 		},
 		{
 			files:  map[string]string{"bar.yml": testRuleBody},
-			finder: discovery.NewGlobFinder([]string{"foo/*"}, discovery.NewPathFilter(nil, nil, nil)),
+			finder: discovery.NewGlobFinder([]string{"foo/*"}, git.NewPathFilter(nil, nil, nil)),
 			err:    "no matching files",
 		},
 		{
 			files:  map[string]string{"bar.yml": testRuleBody},
-			finder: discovery.NewGlobFinder([]string{"*"}, discovery.NewPathFilter(nil, nil, []*regexp.Regexp{regexp.MustCompile(".*")})),
+			finder: discovery.NewGlobFinder([]string{"*"}, git.NewPathFilter(nil, nil, []*regexp.Regexp{regexp.MustCompile(".*")})),
 			entries: []discovery.Entry{
 				{
 					State:         discovery.Noop,
@@ -82,7 +82,7 @@ func TestGlobPathFinder(t *testing.T) {
 		},
 		{
 			files:  map[string]string{"foo/bar.yml": testRuleBody + "\n\n# pint file/owner alice\n"},
-			finder: discovery.NewGlobFinder([]string{"*"}, discovery.NewPathFilter(nil, nil, []*regexp.Regexp{regexp.MustCompile(".*")})),
+			finder: discovery.NewGlobFinder([]string{"*"}, git.NewPathFilter(nil, nil, []*regexp.Regexp{regexp.MustCompile(".*")})),
 			entries: []discovery.Entry{
 				{
 					State:         discovery.Noop,
@@ -96,7 +96,7 @@ func TestGlobPathFinder(t *testing.T) {
 		},
 		{
 			files:  map[string]string{"bar.yml": testRuleBody},
-			finder: discovery.NewGlobFinder([]string{"*"}, discovery.NewPathFilter(nil, nil, nil)),
+			finder: discovery.NewGlobFinder([]string{"*"}, git.NewPathFilter(nil, nil, nil)),
 			entries: []discovery.Entry{
 				{
 					State:         discovery.Noop,
@@ -110,7 +110,7 @@ func TestGlobPathFinder(t *testing.T) {
 		},
 		{
 			files:  map[string]string{"bar.yml": "record:::{}\n  expr: sum(foo)\n\n# pint file/owner bob\n"},
-			finder: discovery.NewGlobFinder([]string{"*"}, discovery.NewPathFilter(nil, nil, []*regexp.Regexp{regexp.MustCompile(".*")})),
+			finder: discovery.NewGlobFinder([]string{"*"}, git.NewPathFilter(nil, nil, []*regexp.Regexp{regexp.MustCompile(".*")})),
 			entries: []discovery.Entry{
 				{
 					State:         discovery.Noop,
@@ -125,7 +125,7 @@ func TestGlobPathFinder(t *testing.T) {
 		{
 			files:    map[string]string{"bar.yml": testRuleBody},
 			symlinks: map[string]string{"link.yml": "bar.yml"},
-			finder:   discovery.NewGlobFinder([]string{"*"}, discovery.NewPathFilter(nil, nil, nil)),
+			finder:   discovery.NewGlobFinder([]string{"*"}, git.NewPathFilter(nil, nil, nil)),
 			entries: []discovery.Entry{
 				{
 					State:         discovery.Noop,
@@ -151,7 +151,7 @@ func TestGlobPathFinder(t *testing.T) {
 				"b/link.yml":   "../a/bar.yml",
 				"b/c/link.yml": "../../a/bar.yml",
 			},
-			finder: discovery.NewGlobFinder([]string{"*"}, discovery.NewPathFilter(nil, nil, nil)),
+			finder: discovery.NewGlobFinder([]string{"*"}, git.NewPathFilter(nil, nil, nil)),
 			entries: []discovery.Entry{
 				{
 					State:         discovery.Noop,
@@ -185,22 +185,22 @@ func TestGlobPathFinder(t *testing.T) {
 				"b/link.yml":   "../a/bar.yml",
 				"b/c/link.yml": "../a/bar.yml",
 			},
-			finder: discovery.NewGlobFinder([]string{"*"}, discovery.NewPathFilter(nil, nil, nil)),
-			err:    "lstat b/a: no such file or directory",
+			finder: discovery.NewGlobFinder([]string{"*"}, git.NewPathFilter(nil, nil, nil)),
+			err:    "b/c/link.yml is a symlink but target file cannot be evaluated: lstat b/a: no such file or directory",
 		},
 		{
 			files: map[string]string{"a/bar.yml": "xxx:\n"},
 			symlinks: map[string]string{
 				"b/c/link.yml": "../../a/bar.yml",
 			},
-			finder: discovery.NewGlobFinder([]string{"*"}, discovery.NewPathFilter(nil, nil, []*regexp.Regexp{regexp.MustCompile(".*")})),
+			finder: discovery.NewGlobFinder([]string{"*"}, git.NewPathFilter(nil, nil, []*regexp.Regexp{regexp.MustCompile(".*")})),
 		},
 		{
 			files: map[string]string{"a/bar.yml": "xxx:\nyyy:\n"},
 			symlinks: map[string]string{
 				"b/c/link.yml": "../../a/bar.yml",
 			},
-			finder: discovery.NewGlobFinder([]string{"*"}, discovery.NewPathFilter(nil, nil, nil)),
+			finder: discovery.NewGlobFinder([]string{"*"}, git.NewPathFilter(nil, nil, nil)),
 			entries: []discovery.Entry{
 				{
 					State:         discovery.Noop,
@@ -225,21 +225,21 @@ func TestGlobPathFinder(t *testing.T) {
 			symlinks: map[string]string{
 				"b/c/link.yml": "../../a/bar.yml",
 			},
-			finder: discovery.NewGlobFinder([]string{"*"}, discovery.NewPathFilter(nil, nil, []*regexp.Regexp{regexp.MustCompile(".*")})),
+			finder: discovery.NewGlobFinder([]string{"*"}, git.NewPathFilter(nil, nil, []*regexp.Regexp{regexp.MustCompile(".*")})),
 		},
 		{
 			files: map[string]string{"a/bar.yml": "xxx:\nyyy:\n"},
 			symlinks: map[string]string{
 				"b/c/d": "../../a",
 			},
-			finder: discovery.NewGlobFinder([]string{"*"}, discovery.NewPathFilter(nil, nil, []*regexp.Regexp{regexp.MustCompile(".*")})),
+			finder: discovery.NewGlobFinder([]string{"*"}, git.NewPathFilter(nil, nil, []*regexp.Regexp{regexp.MustCompile(".*")})),
 		},
 		{
 			files: map[string]string{"a/bar.yml": testRuleBody},
 			symlinks: map[string]string{
 				"b/c/link.yml": "../../a/bar.yml",
 			},
-			finder: discovery.NewGlobFinder([]string{"*"}, discovery.NewPathFilter(nil, nil, nil)),
+			finder: discovery.NewGlobFinder([]string{"*"}, git.NewPathFilter(nil, nil, nil)),
 			entries: []discovery.Entry{
 				{
 					State:         discovery.Noop,
@@ -263,8 +263,8 @@ func TestGlobPathFinder(t *testing.T) {
 			symlinks: map[string]string{
 				"input.yml": "/xx/ccc/fdd",
 			},
-			finder: discovery.NewGlobFinder([]string{"*"}, discovery.NewPathFilter(nil, nil, nil)),
-			err:    "stat input.yml: no such file or directory",
+			finder: discovery.NewGlobFinder([]string{"*"}, git.NewPathFilter(nil, nil, nil)),
+			err:    "input.yml is a symlink but target file cannot be evaluated: lstat /xx: no such file or directory",
 		},
 	}